@@ -1,26 +1,29 @@
 {
-  "homepage": "http://ameet2r.github.io",
-  "name": "ameet2r.github.io",
-  "version": "1.0.0",
+  "name": "ameet-toor-portfolio",
+  "version": "2.0.0",
   "private": true,
+  "description": "Professional portfolio for Ameet Toor - Full-Stack Software Engineer",
+  "author": "Ameet Toor",
+  "keywords": ["portfolio", "freelance", "software-engineer", "full-stack", "react"],
   "dependencies": {
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/icons": "^4.11.2",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "^5.0.0",
+    "@mui/material": "^5.14.0",
+    "@mui/icons-material": "^5.14.0",
+    "@mui/styles": "^5.14.0",
+    "@emotion/react": "^11.11.0",
+    "@emotion/styled": "^11.11.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.45.4",
+    "resend": "^1.1.0",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "test": "vitest"
   },
   "eslintConfig": {
     "extends": [
@@ -41,6 +44,13 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^3.2.0"
+    "@testing-library/jest-dom": "^5.14.1",
+    "@testing-library/react": "^11.2.7",
+    "@testing-library/user-event": "^12.8.3",
+    "@vitejs/plugin-react": "^4.0.0",
+    "gh-pages": "^3.2.0",
+    "jsdom": "^22.0.0",
+    "vite": "^4.3.0",
+    "vitest": "^0.31.0"
   }
 }
