@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
-import { ThemeProvider as MThemeProvider } from '@material-ui/core/styles';
-import CssBaseline from '@material-ui/core/CssBaseline';
+import { ThemeProvider as MThemeProvider, StyledEngineProvider } from '@mui/material/styles';
+import CssBaseline from '@mui/material/CssBaseline';
 import dark from './Dark';
 import light from './Light';
 
@@ -34,9 +34,11 @@ export default function ThemeProvider(props) {
     const isDarkTheme = useThemeDetector();
 
     return (
-        <MThemeProvider theme={isDarkTheme ? dark : light}>
-            <CssBaseline />
-            {props.children}
-        </MThemeProvider>
+        <StyledEngineProvider injectFirst>
+            <MThemeProvider theme={isDarkTheme ? dark : light}>
+                <CssBaseline />
+                {props.children}
+            </MThemeProvider>
+        </StyledEngineProvider>
     )
 };
